script de récupération

+++ installation.sh
@@ -1,4 +1,4 @@
-# #!/bin/bash
+#!/bin/bash
 
 echo "modification de la memoire..."
 
@@ -7,6 +7,18 @@ chmod 600 /swapfile1
 mkswap /swapfile1
 swapon /swapfile1
 
+REP_SWAPFILE=""
+# Vérifie si la ligne existe déjà dans /etc/fstab
+if ! grep -q "/swapfile1" /etc/fstab; then
+    echo "/swapfile1 none swap sw 0 0" | sudo tee -a /etc/fstab > /dev/null
+    
+    REP_SWAPFILE="Ligne ajoutée à /etc/fstab"
+    echo "Ligne ajoutée à /etc/fstab"
+else
+    REP_SWAPFILE="SwapFile : La ligne existe déjà, elle n'a pas été ajoutée."
+    echo "La ligne existe déjà dans /etc/fstab"
+fi
+
 echo "update du vps..."
 sudo apt update -y
 sudo apt install -y vim
@@ -14,13 +26,21 @@ sudo apt install -y vim
 echo "ajout d'utilisateur..."
 
 # # Créer l’utilisateur avec son mot de passe passé en argument
-sudo useradd -m -s /bin/bash -G sudo maxence
-echo "maxence:$1" | sudo chpasswd
 
-echo "modification du service SSH..."
-sudo cp /etc/ssh/sshd_config /etc/ssh/sshd_config.bak
+REP_USER=""
 
-sudo sed -i 's/^#\?PermitRootLogin.*/PermitRootLogin no/' /etc/ssh/sshd_config
+if sudo useradd -m -s /bin/bash -G sudo maxence; then
+    echo "maxence:$1" | sudo chpasswd
+    echo "modification du service SSH..."
+    sudo cp /etc/ssh/sshd_config /etc/ssh/sshd_config.bak
+    sudo sed -i 's/^#\?PermitRootLogin.*/PermitRootLogin no/' /etc/ssh/sshd_config
+
+    REP_USER="l'utilisateur à bien été crée"
+else
+    REP_USER="l'utilisateur était déjà crée ou une erreur est survenue"
+fi
+
+echo "$REP_USER"
 
 echo "redémarage du service ssh..."
 sudo systemctl restart ssh
@@ -30,37 +50,84 @@ sudo systemctl restart ssh
 ###########################
 
 echo "mise en place du stack lamp..."
-sudo apt install apache2 -y
-sudo apt install php -y
-sudo apt install mariadb-server -y
 
-echo "sécurisation de mysql..."
+REP_APACHE=""
+REP_PHP=""
+REP_MARIADB=""
 
-sudo mysql -e "DELETE FROM mysql.user WHERE User='';"
-sudo mysql -e "DROP DATABASE IF EXISTS test;"
-sudo mysql -e "DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%';"
-sudo mysql -e "UPDATE mysql.user SET Host='localhost' WHERE User='root';"
-sudo mysql -e "FLUSH PRIVILEGES;"
+if sudo apt install apache2 -y; then
+    REP_APACHE="apache installé avec succès";
+else
+    REP_APACHE="une erreur est survenue lors de  l'installation de apache2";
+fi
 
-echo "phpmyadmin phpmyadmin/reconfigure-webserver multiselect apache2" | sudo debconf-set-selections
-echo "phpmyadmin phpmyadmin/dbconfig-install boolean true" | sudo debconf-set-selections
-echo "phpmyadmin phpmyadmin/mysql/admin-pass password $1" | sudo debconf-set-selections
-echo "phpmyadmin phpmyadmin/mysql/app-pass password $1" | sudo debconf-set-selections
-echo "phpmyadmin phpmyadmin/app-password-confirm password $1" | sudo debconf-set-selections
+if sudo apt install php -y; then
+    REP_PHP="php installé avec succès";
+else
+    REP_PHP="une erreur est survenue lors de  l'installation de PHP";
+fi
 
-sudo apt install -y phpmyadmin
+if sudo apt install mariadb-server -y; then
+    REP_MARIADB="mariadb installé avec succès";
+else
+    REP_MARIADB="maria db a rencontré une erreur lors de l'installation";
+fi
 
-echo "création d'un utilisateur..."
-# Variables (modifiable facilement)
-DB_USER="maxence"
+echo "$REP_APACHE"
+echo "$REP_PHP"
+echo "$REP_MARIADB"
 
-# Création de l'utilisateur MySQL et attribution des droits
-sudo mysql -e "CREATE USER '${DB_USER}'@'localhost' IDENTIFIED BY '$1';"
-sudo mysql -e "GRANT ALL PRIVILEGES ON *.* TO '${DB_USER}'@'localhost';"
-sudo mysql -e "FLUSH PRIVILEGES;"
+REP_SECURE=""
+REP_PHPMYADMIN=""
+
+if dpkg -l | grep -q mariadb-server; then
+    echo "MariaDB est installée, sécurisation en cours..."
+    sudo systemctl start mariadb
+    echo "sécurisation de mysql..."
+
+    if \
+    sudo mysql -e "DELETE FROM mysql.user WHERE User='';" && \
+    sudo mysql -e "DROP DATABASE IF EXISTS test;" && \
+    sudo mysql -e "DELETE FROM mysql.db WHERE Db='test' OR Db='test\\_%';" && \
+    sudo mysql -e "UPDATE mysql.user SET Host='localhost' WHERE User='root';" && \
+    sudo mysql -e "FLUSH PRIVILEGES;"; then
+
+        REP_SECURE="Sécurisation MariaDB effectuée avec succès !"
+    else
+        REP_SECURE="Une erreur est survenue lors de la sécurisation de MariaDB"
+    fi
+
+    echo "$REP_SECURE"
+
+    echo "phpmyadmin phpmyadmin/reconfigure-webserver multiselect apache2" | sudo debconf-set-selections
+    echo "phpmyadmin phpmyadmin/dbconfig-install boolean true" | sudo debconf-set-selections
+    echo "phpmyadmin phpmyadmin/mysql/admin-pass password $1" | sudo debconf-set-selections
+    echo "phpmyadmin phpmyadmin/mysql/app-pass password $1" | sudo debconf-set-selections
+    echo "phpmyadmin phpmyadmin/app-password-confirm password $1" | sudo debconf-set-selections
+
+    if sudo apt install -y phpmyadmin; then
+        REP_PHPMYADMIN="phpmyadmin a été installé"
+
+        echo "création d'un utilisateur..."
+        # Variables (modifiable facilement)
+        DB_USER="maxence"
+
+        # Création de l'utilisateur MySQL et attribution des droits
+        sudo mysql -e "CREATE USER '${DB_USER}'@'localhost' IDENTIFIED BY '$1';"
+        sudo mysql -e "GRANT ALL PRIVILEGES ON *.* TO '${DB_USER}'@'localhost';"
+        sudo mysql -e "FLUSH PRIVILEGES;"
+    else
+        REP_PHPMYADMIN="phpmyadmin a rencontré une erreur lors de l'installation"
+    fi
+
+    echo "$REP_PHPMYADMIN"
+else
+    echo "MariaDB n'est pas installée. Installation requise avant la sécurisation et l'installation de PHPMYADMIN."
+fi
 
 sudo a2enmod rewrite
 sudo systemctl restart apache2
+
 
 echo "récupération des dossiers..."
 cd /var/www/html
@@ -74,12 +141,32 @@ cd /etc/apache2/sites-available
 wget "https://raw.githubusercontent.com/maxsanch/SAE501/refs/heads/main/perso-ssl.conf"
 
 wget "https://raw.githubusercontent.com/maxsanch/SAE501/refs/heads/main/perso.conf"
-sudo a2ensite perso.conf
 
-sudo a2enmod ssl
+REP_SITE=""
+
+if sudo a2ensite perso.conf; then
+    REP_SITE="le site a bien été configuré"
+else
+    REP_SITE="la configuration du site a echouée"
+fi
+
+REP_SSL=""
+
+if sudo a2enmod ssl; then 
+    REP_SSL="le module SSL a bien été activé"
+else
+    REP_SSL="une erreur est survenue lors de l'activation du module SSL"
+fi
 
 echo "activation du site..."
-sudo a2ensite perso-ssl.conf
+
+REP_SITESSL=""
+
+if sudo a2ensite perso-ssl.conf; then
+    REP_SITESSL="le site est opérationnel avec SSL !"
+else
+    REP_SITESSL="la configuration du site SSL n'a pas fonctionnée, veuillez le faire manuellement."
+fi
 
 echo "redemarage de apache..."
 sudo systemctl reload apache2
@@ -87,20 +174,63 @@ sudo systemctl reload apache2
 cd /var/www/html/perso
 
 echo "recuperation du fichier prestashop..."
-wget https://assets.prestashop3.com/dst/edition/corporate/9.0.0-1.0/prestashop_edition_classic_version_9.0.0-1.0.zip?source=docker
-mv 'prestashop_edition_classic_version_9.0.0-1.0.zip?source=docker' prestashop.zip
-echo "recup ok"
+
+REP_GETPRESTA=""
+
+if wget https://assets.prestashop3.com/dst/edition/corporate/9.0.0-1.0/prestashop_edition_classic_version_9.0.0-1.0.zip?source=docker; then
+    REP_GETPRESTA="prestashop recupéré"
+else
+    REP_GETPRESTA="La récupération du prestashop a échouée"
+fi
+
+echo "$REP_GETPRESTA"
+
+REP_MOVE=""
+
+if mv 'prestashop_edition_classic_version_9.0.0-1.0.zip?source=docker' prestashop.zip; then
+    REP_MOVE="le déplacement du fichier a été effectué"
+else
+    REP_MOVE="le déplacement du fichier a échoué"
+fi
+
+echo "$REP_MOVE"
 
 sudo apt update
-sudo apt install unzip
 
-unzip -o prestashop.zip
+REP_UNZIP=""
+
+if sudo apt install unzip; then
+    REP_UNZIP="installation de UNZIP effectuée"
+else
+    REP_UNZIP="m'installation de UNZIP a échouée"
+fi
+
+echo "$REP_UNZIP"
+
+REP_UNZIPDONE=""
+
+if unzip -o prestashop.zip; then
+    REP_UNZIPDONE="dézippage de prestashop effectué"
+else
+    REP_UNZIPDONE="le dezippage de prestashop a echoué"
+fi
+
+echo "$REP_UNZIPDONE"
 
 chown www-data *
 chown www-data .
 
 echo "dezippage du projet..."
-unzip -o prestashop.zip
+
+REP_UNZIPDONETWO=""
+
+if unzip -o prestashop.zip; then
+    REP_UNZIPDONETWO="le second dezippage a été effectué"
+else
+    REP_UNZIPDONETWO="le second dezippage a échoué"
+fi
+
+echo "$REP_UNZIPDONETWO"
 
 sudo apt update
 sudo apt install -y unzip php-intl
@@ -114,7 +244,14 @@ sudo find /var/www/html/perso -type f -exec chmod 644 {} \;
 echo "recherche index cli..."
 
 cd /var/www/html/perso/install
-php index_cli.php --domain=www.prestashopexo.com --db_server=127.0.0.1 --db_name=SAEShop --db_user=maxence  --db_password=$1  --prefix=myshop_ --email=maxence.sanchez05@gmail.com --password=$1
+
+REP_CLI=""
+
+if php index_cli.php --domain=www.prestashopexo.com --db_server=127.0.0.1 --db_name=SAEShop --db_user=maxence  --db_password=$1  --prefix=myshop_ --email=maxence.sanchez05@gmail.com --password=$1; then
+    REP_CLI="installation CLi effectuée"
+else
+    REP_CLI="l'installation CLI a échouée."
+fi
 
 cd /var/www/html/perso
 
@@ -130,3 +267,19 @@ sudo mysql -u $DB_USER -p$1 -e "UPDATE ${DB_NAME}.${DB_PREFIX}configuration SET 
 sudo systemctl restart apache2
 
 echo "-- normalement, c'est bon !--"
+echo "$REP_SWAPFILE"
+echo "$REP_USER"
+echo "$REP_APACHE"
+echo "$REP_PHP"
+echo "$REP_MARIADB"
+echo "$REP_SECURE"
+echo "$REP_PHPMYADMIN"
+echo "$REP_SITE"
+echo "$REP_SSL"
+echo "$REP_SITESSL"
+echo "$REP_GETPRESTA"
+echo "$REP_MOVE"
+echo "$REP_UNZIP"
+echo "$REP_UNZIPDONE"
+echo "$REP_UNZIPDONETWO"
+echo "$REP_CLI"
